# [aten] Pass std::function to thread_pool by value, instead of const ref.