--- a/c10/core/thread_pool.h
+++ b/c10/core/thread_pool.h
@@ -17,7 +17,7 @@ namespace c10 {
 // TODO: move this to C10 and make it C10_API
 class C10_API TaskThreadPoolBase {
  public:
-  virtual void run(const std::function<void()>& func) = 0;
+  virtual void run(std::function<void()> func) = 0;
 
   virtual size_t size() const = 0;
 
@@ -49,10 +49,10 @@ class C10_API ThreadPool : public c10::TaskThreadPoolBase {
     const std::function<void()> no_id;
     const std::function<void(std::size_t)> with_id;
 
-    explicit task_element_t(const std::function<void()>& f)
-        : run_with_id(false), no_id(f), with_id(nullptr) {}
-    explicit task_element_t(const std::function<void(std::size_t)>& f)
-        : run_with_id(true), no_id(nullptr), with_id(f) {}
+    explicit task_element_t(std::function<void()> f)
+      : run_with_id(false), no_id(std::move(f)), with_id(nullptr) {}
+    explicit task_element_t(std::function<void(std::size_t)> f)
+      : run_with_id(true), no_id(nullptr), with_id(std::move(f)) {}
   };
 
   std::queue<task_element_t> tasks_;
@@ -82,7 +82,7 @@ class C10_API ThreadPool : public c10::TaskThreadPoolBase {
 
   bool inThreadPool() const override;
 
-  void run(const std::function<void()>& func) override;
+  void run(std::function<void()> func) override;
 
   template <typename Task>
   void runTaskWithID(Task task) {
@@ -90,8 +90,7 @@ class C10_API ThreadPool : public c10::TaskThreadPoolBase {
 
     // Set task and signal condition variable so that a worker thread will
     // wake up and use the task.
-    tasks_.push(
-        task_element_t(static_cast<std::function<void(std::size_t)>>(task)));
+    tasks_.emplace(static_cast<std::function<void(std::size_t)>>(task));
     complete_ = false;
     condition_.notify_one();
   }