--- a/test/cpp/api/CMakeLists.txt
+++ b/test/cpp/api/CMakeLists.txt
@@ -73,3 +73,7 @@ if(INSTALL_TEST)
     install(FILES $<TARGET_PDB_FILE:test_api> DESTINATION bin OPTIONAL)
   endif()
 endif()
+
+add_executable(parallel_benchmark ${TORCH_API_TEST_DIR}/parallel_benchmark.cpp)
+target_include_directories(parallel_benchmark PRIVATE ${ATen_CPU_INCLUDE})
+target_link_libraries(parallel_benchmark PRIVATE torch)