Stack from [ghstack](https://github.com/ezyang/ghstack):
* **#37681 [aten] Pass std::function<> to thread_pool by value, instead of const ref.**

By passing by value, we can std::move, and avoid unnecessarily copying
args that are part of any std::function/lambda state (e.g. in the jit
interpreter, there is a std::vector<> stack passed in the
InterpreterContinuation)

This makes the api also consistent with e.g. folly. Added a minor
at::launch() benchmark to test/cpp/, the difference is mostly
noticeable when copying the std::function<> internal args is
non-trivial.

Benchmarks pre/post (min over ~5 runs) 
NoData:      5.81 us -> 5.63 us (-3.2%)                                                                                                 
WithData(0): 6.67 us -> 5.88 us (-11.8%)                                                                                                
WithData(4): 6.98 us -> 6.51 us (-6.7%)                                                                                                 
WithData(256): 9.44 us -> 7.89 (-16.5%)                                                                                                 
                                                       
Differential Revision: [D21355148](https://our.internmc.facebook.com/intern/diff/D21355148/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D21355148/)!